feat(newOnboarding): improve animation
