+++ src/components/organisms/OnBoarding/Steps/index.ts
@@ -1,4 +1,4 @@
-import { CloseFunction } from '../../index';
+import { CloseFunction } from '../index';
 
 export interface StepTypes {
   next?: () => void;
@@ -8,3 +8,4 @@ export interface StepTypes {
 
 export { default as Step1 } from './Step1';
 export { default as Step2 } from './Step2';
+export { default as Step3 } from './Step3';
